# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sarkari Exams API",
    "version": "1.0.0",
    "description": "Shared API contract for frontend/backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/subscriptions": {
      "post": {
        "summary": "Create or update subscription",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/subscriptions/unsubscribe": {
      "get": {
        "summary": "Unsubscribe",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unsubscribe response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/subscriptions/verify": {
      "get": {
        "summary": "Verify subscription",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verify response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "SubscriptionRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "frequency": {
            "type": "string",
            "enum": [
              "instant",
              "daily",
              "weekly"
            ]
          }
        },
        "required": [
          "email"
        ]
      },
      "SubscriptionResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "verified": {
                "type": "boolean"
              }
            },
            "required": [
              "verified"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}